migrate to new implementation based on the documentation
